Fix windows CI
