+++ ehrapy/preprocessing/__init__.py
@@ -22,3 +22,4 @@ from ehrapy.preprocessing._normalization import (
 from ehrapy.preprocessing._quality_control import qc_lab_measurements, qc_metrics
 from ehrapy.preprocessing._scanpy_pp_api import *  # noqa: E402,F403
 from ehrapy.preprocessing.encoding._encode import encode, undo_encoding
+from ehrapy.preprocessing.highly_variable_features import highly_variable_features
